clang warnings

+++ adb.c
@@ -250,7 +250,7 @@ static const char *adb_init(void *frontend_handle, void **backend_handle,
 #           define ADB_SHELL_HOST_MAX_LEN (sizeof(sendbuf)-4-ADB_SHELL_SERIAL_PREFIX_LEN)
             if (len > ADB_SHELL_HOST_MAX_LEN)
                 len = ADB_SHELL_HOST_MAX_LEN;
-            sprintf(sendbuf,"%04lx" ADB_SHELL_SERIAL_PREFIX, len+ADB_SHELL_SERIAL_PREFIX_LEN);
+            sprintf(sendbuf,"%04lx" ADB_SHELL_SERIAL_PREFIX, (unsigned long)(len+ADB_SHELL_SERIAL_PREFIX_LEN));
             memcpy(sendbuf+4+ADB_SHELL_SERIAL_PREFIX_LEN, host, len);
             write_hello(sendbuf, len+4+ADB_SHELL_SERIAL_PREFIX_LEN);
         }
